+++ plugins/help.py
@@ -6,7 +6,8 @@ import re
 class pluginHelp(urusai_plugin.MucMessage):
     """
     Information about available modules in this MUC and their usability.
-    The data is being given from classes' docstrings
+    The data is being given from classes' docstrings.
+    Warning: don't move this plugin under any subdirectories to keep it working fine.
     Usage:
         "help" for available modules listing
         "help <MODULE>" for the details about specific module